Move elsterify methods to separate file
